+++ test/checks.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 
-rep_HD_list="sod-iso sod MachReflection ViscousFlowPastCylinder ViscousDisk FargoPlanet"
-rep_MHD_list="sod-iso sod AxisFluxTube AmbipolarCshock HallWhistler FargoMHDSpherical OrszagTang OrszagTang3D"
+rep_HD_list="sod-iso sod MachReflection ViscousFlowPastCylinder ViscousDisk FargoPlanet ShearingBox"
+rep_MHD_list="sod-iso sod AxisFluxTube AmbipolarCshock HallWhistler FargoMHDSpherical ShearingBox OrszagTang OrszagTang3D"
 
 # refer to the parent dir of this file, wherever this is called from
 # a python equivalent is e.g.
@@ -40,26 +40,28 @@ for rep in $rep_HD_list; do
     echo "***********************************************"
     echo "Configuring  $rep"
     echo "***********************************************"
+    rm -f CMakeCache.txt
     def_files=$(ls definitions*.hpp)
     for def in $def_files; do
-        python3 $IDEFIX_DIR/configure.py $options -defs=$def
+
+        cmake $IDEFIX_DIR $options -DIdefix_DEFS=$def || { echo "!!!! HD $rep failed during configuration"; exit 1; }
         echo "***********************************************"
         echo "Making  $rep with $def"
         echo "***********************************************"
-        make clean; make -j 4
+        make clean; make -j 4 || { echo "!!!! HD $rep failed during compilation with $def"; exit 1; }
 
         ini_files=$(ls *.ini)
         for ini in $ini_files; do
             echo "***********************************************"
             echo "Running  $rep with $ini"
             echo "***********************************************"
-            ./idefix -i $ini
+            ./idefix -i $ini || { echo "!!!! HD $rep failed running with $def and $ini"; exit 1; }
 
             cd python
             echo "***********************************************"
             echo "Testing  $rep with $ini and $def"
             echo "***********************************************"
-            python3 testidefix.py -noplot
+            python3 testidefix.py -noplot || { echo "!!!! HD $rep failed validation with $def and $ini"; exit 1; }
             cd ..
         done
         make clean
@@ -73,28 +75,45 @@ for rep in $rep_MHD_list; do
     echo "***********************************************"
     echo "Configuring  $rep"
     echo "***********************************************"
+    rm -f CMakeCache.txt
     def_files=$(ls definitions*.hpp)
     for def in $def_files; do
-        python3 $IDEFIX_DIR/configure.py -mhd $options -defs=$def
+        cmake $IDEFIX_DIR $options -DIdefix_DEFS=$def -DIdefix_MHD=ON|| { echo "!!!! MHD $rep failed during configuration"; exit 1; }
         echo "***********************************************"
         echo "Making  $rep with $def"
         echo "***********************************************"
-        make clean; make -j 4
+        make clean; make -j 4 || { echo "!!!! MHD $rep failed during compilation with $def"; exit 1; }
 
         ini_files=$(ls *.ini)
         for ini in $ini_files; do
             echo "***********************************************"
             echo "Running  $rep with $ini"
             echo "***********************************************"
-            ./idefix -i $ini
+            ./idefix -i $ini || { echo "!!!! MHD $rep failed running with $def and $ini"; exit 1; }
 
             cd python
             echo "***********************************************"
             echo "Testing  $rep with $ini and $def"
             echo "***********************************************"
-            python3 testidefix.py -noplot
+            python3 testidefix.py -noplot || { echo "!!!! MHD $rep failed validation with $def and $ini"; exit 1; }
             cd ..
         done
     done
     cd $TEST_DIR
 done
+
+# Test restart functions with OT3D which have generated a dump during the first pass
+rep=OrszagTang3D
+cd $TEST_DIR/MHD/$rep
+# remove generated vtk from previous run
+rm *.vtk
+echo "***********************************************"
+echo "Running  $rep with restart dump # 1"
+echo "***********************************************"
+./idefix -restart 1 || { echo "!!!! MHD $rep failed running restart dump validation"; exit 1; }
+cd python
+echo "***********************************************"
+echo "Testing  $rep with restart dump # 1"
+echo "***********************************************"
+python3 testidefix.py -noplot || { echo "!!!! MHD $rep failed checking restart dump validation"; exit 1; }
+cd ..