+++ test/checks_mpi.sh
@@ -41,7 +41,8 @@ for rep in $rep_HD_2D_mpi_list; do
     echo "***********************************************"
     echo "Configuring  $rep"
     echo "***********************************************"
-    python3 $IDEFIX_DIR/configure.py -mpi $options
+    rm -f CMakeCache.txt
+    cmake $IDEFIX_DIR -DIdefix_MPI=ON $options
     echo "***********************************************"
     echo "Making  $rep"
     echo "***********************************************"
@@ -71,7 +72,8 @@ for rep in $rep_MHD_2D_mpi_list; do
     echo "***********************************************"
     echo "Configuring  $rep"
     echo "***********************************************"
-    python3 $IDEFIX_DIR/configure.py -mhd -mpi $options
+    rm -f CMakeCache.txt
+    cmake $IDEFIX_DIR -DIdefix_MHD=ON -DIdefix_MPI=ON $options
     echo "***********************************************"
     echo "Making  $rep"
     echo "***********************************************"
@@ -101,7 +103,8 @@ for rep in $rep_MHD_3D_mpi_list; do
     echo "***********************************************"
     echo "Configuring  $rep"
     echo "***********************************************"
-    python3 $IDEFIX_DIR/configure.py -mhd -mpi $options
+    rm -f CMakeCache.txt
+    cmake $IDEFIX_DIR -DIdefix_MHD=ON -DIdefix_MPI=ON $options
     echo "***********************************************"
     echo "Making  $rep"
     echo "***********************************************"
@@ -124,3 +127,19 @@ for rep in $rep_MHD_3D_mpi_list; do
 
     cd $TEST_DIR
 done
+
+# Test restart functions with OT3D which have generated a dump during the first pass
+rep=OrszagTang3D
+cd $TEST_DIR/MHD/$rep
+# remove generated vtk from previous run
+rm *.vtk
+echo "***********************************************"
+echo "Running  $rep with restart dump # 1"
+echo "***********************************************"
+mpirun -np 8 ./idefix -restart 1 -dec 2 2 2 || { echo "!!!! MHD $rep failed running restart dump validation"; exit 1; }
+cd python
+echo "***********************************************"
+echo "Testing  $rep with restart dump # 1"
+echo "***********************************************"
+python3 testidefix.py -noplot || { echo "!!!! MHD $rep failed checking restart dump validation"; exit 1; }
+cd ..