+++ test/test_configure.py
@@ -3,6 +3,7 @@
 # see https://docs.pytest.org/en/stable/usage.html#cmdline
 import os
 from pathlib import Path
+import textwrap
 import pytest
 
 from configure import main, GPU_ARCHS, CPU_ARCHS  # noqa: E402
@@ -20,7 +21,12 @@ def in_tmp_dir(tmp_path):
         os.chdir(previous_dir)
 
 
-@pytest.mark.usefixtures("in_tmp_dir")
+@pytest.fixture()
+def no_config_file(monkeypatch, tmp_path):
+    monkeypatch.setenv("XDG_CONFIG_HOME", str(tmp_path))
+
+
+@pytest.mark.usefixtures("in_tmp_dir", "no_config_file")
 def test_main_no_idefix(capsys, monkeypatch, tmp_path):
     monkeypatch.delenv("IDEFIX_DIR", raising=False)
     ret = main([])
@@ -32,7 +38,7 @@ def test_main_no_idefix(capsys, monkeypatch, tmp_path):
     assert not Path("Makefile").is_file()
 
 
-@pytest.mark.usefixtures("in_tmp_dir")
+@pytest.mark.usefixtures("in_tmp_dir", "no_config_file")
 def test_parse_archs_overload(capsys, monkeypatch):
     monkeypatch.setenv("IDEFIX_DIR", IDEFIX_DIR)
     ret = main(["-arch", "BDW", "Kepler30", "HSW"])
@@ -44,7 +50,7 @@ def test_parse_archs_overload(capsys, monkeypatch):
     assert not Path("Makefile").is_file()
 
 
-@pytest.mark.usefixtures("in_tmp_dir")
+@pytest.mark.usefixtures("in_tmp_dir", "no_config_file")
 def test_parse_archs_two_cpus(capsys, monkeypatch):
     monkeypatch.setenv("IDEFIX_DIR", IDEFIX_DIR)
     ret = main(["-arch", "BDW", "HSW"])
@@ -56,7 +62,7 @@ def test_parse_archs_two_cpus(capsys, monkeypatch):
     assert not Path("Makefile").is_file()
 
 
-@pytest.mark.usefixtures("in_tmp_dir")
+@pytest.mark.usefixtures("in_tmp_dir", "no_config_file")
 def test_parse_archs_two_gpus(capsys, monkeypatch):
     monkeypatch.setenv("IDEFIX_DIR", IDEFIX_DIR)
     ret = main(["-arch", "Kepler30", "Maxwell50"])
@@ -68,7 +74,7 @@ def test_parse_archs_two_gpus(capsys, monkeypatch):
     assert not Path("Makefile").is_file()
 
 
-@pytest.mark.usefixtures("in_tmp_dir")
+@pytest.mark.usefixtures("in_tmp_dir", "no_config_file")
 def test_main_default_success(capsys, monkeypatch):
     monkeypatch.setenv("IDEFIX_DIR", IDEFIX_DIR)
     ret = main([])
@@ -82,7 +88,7 @@ def test_main_default_success(capsys, monkeypatch):
 
 
 @pytest.mark.parametrize("arch", GPU_ARCHS)
-@pytest.mark.usefixtures("in_tmp_dir")
+@pytest.mark.usefixtures("in_tmp_dir", "no_config_file")
 def test_main_auto_gpu_mode_solo(arch, capsys, monkeypatch):
     monkeypatch.setenv("IDEFIX_DIR", IDEFIX_DIR)
     ret = main(["-arch", arch])
@@ -95,7 +101,7 @@ def test_main_auto_gpu_mode_solo(arch, capsys, monkeypatch):
 
 
 @pytest.mark.parametrize("arch", CPU_ARCHS)
-@pytest.mark.usefixtures("in_tmp_dir")
+@pytest.mark.usefixtures("in_tmp_dir", "no_config_file")
 def test_main_auto_cpu_mode_solo(arch, capsys, monkeypatch):
     monkeypatch.setenv("IDEFIX_DIR", IDEFIX_DIR)
     ret = main(["-arch", arch])
@@ -107,7 +113,7 @@ def test_main_auto_cpu_mode_solo(arch, capsys, monkeypatch):
     assert Path("Makefile").is_file()
 
 
-@pytest.mark.usefixtures("in_tmp_dir")
+@pytest.mark.usefixtures("in_tmp_dir", "no_config_file")
 def test_main_gpu_flag_deprecation(capsys, monkeypatch):
     from configure import _GPU_FLAG_DEPRECATION_MESSAGE
     monkeypatch.setenv("IDEFIX_DIR", IDEFIX_DIR)
@@ -117,4 +123,110 @@ def test_main_gpu_flag_deprecation(capsys, monkeypatch):
     out, err = capsys.readouterr()
     assert "Execution target: GPU" in out
     assert err == "Warning: %s\n" % _GPU_FLAG_DEPRECATION_MESSAGE
+    assert Path("Makefile").is_file()
+
+
+@pytest.fixture()
+def tmp_config_file_gpu(monkeypatch, tmp_path):
+    monkeypatch.setenv("XDG_CONFIG_HOME", str(tmp_path))
+    file = tmp_path / "idefix.cfg"
+    with open(file, "w") as fh:
+        fh.write(
+            textwrap.dedent(
+                """
+                [compilation]
+                GPU = Turing75
+                """,
+            ),
+        )
+    return file
+
+
+@pytest.mark.usefixtures("in_tmp_dir", "tmp_config_file_gpu")
+def test_main_gpu_from_conf_file(capsys, monkeypatch):
+    monkeypatch.setenv("IDEFIX_DIR", IDEFIX_DIR)
+
+    # explicitly request a CPU arch on the command line
+    ret = main(["-arch", "EPYC"])
+    assert ret == 0
+
+    out, err = capsys.readouterr()
+    assert "Execution target: GPU\n" in out
+    assert "Target architecture: Turing75\n" in out
+    assert err == ""
+    assert Path("Makefile").is_file()
+
+
+@pytest.mark.usefixtures("in_tmp_dir", "tmp_config_file_gpu")
+def test_main_arch_from_conf_file(capsys, monkeypatch):
+    monkeypatch.setenv("IDEFIX_DIR", IDEFIX_DIR)
+
+    ret = main([])
+    assert ret == 0
+
+    out, err = capsys.readouterr()
+    assert "Execution target: GPU\n" in out
+    assert "Target architecture: Turing75\n" in out
+    assert err == ""
+    assert Path("Makefile").is_file()
+
+
+@pytest.mark.usefixtures("in_tmp_dir", "tmp_config_file_gpu")
+def test_main_arch_cli_override_conf_file(capsys, monkeypatch):
+    monkeypatch.setenv("IDEFIX_DIR", IDEFIX_DIR)
+
+    ret = main(["-arch", "Volta70"])
+    assert ret == 0
+
+    out, err = capsys.readouterr()
+    assert "Execution target: GPU\n" in out
+    assert "Target architecture: Volta70\n" in out
+    assert err == ""
+    assert Path("Makefile").is_file()
+
+
+@pytest.mark.usefixtures("in_tmp_dir", "tmp_config_file_gpu")
+def test_main_arch_local_override_global_conf_file(capsys, monkeypatch):
+    monkeypatch.setenv("IDEFIX_DIR", IDEFIX_DIR)
+
+    with open("idefix.cfg", "w") as fh:
+        fh.write(
+            textwrap.dedent(
+                """
+                [compilation]
+                GPU = Pascal60
+                """,
+            ),
+        )
+
+    ret = main([])
+    assert ret == 0
+
+    out, err = capsys.readouterr()
+    assert "Execution target: GPU\n" in out
+    assert "Target architecture: Pascal60\n" in out
+    assert err == ""
+    assert Path("Makefile").is_file()
+
+
+@pytest.mark.usefixtures("in_tmp_dir")
+def test_main_set_compiler_from_conf_file(capsys, monkeypatch):
+    monkeypatch.setenv("IDEFIX_DIR", IDEFIX_DIR)
+
+    with open("idefix.cfg", "w") as fh:
+        fh.write(
+            textwrap.dedent(
+                """
+                [compilation]
+                CXX = MY_COMPILER
+                """,
+            ),
+        )
+
+    ret = main([])
+    assert ret == 0
+
+    out, err = capsys.readouterr()
+    assert "Compiler: MY_COMPILER\n" in out
+    assert err == ""
     assert Path("Makefile").is_file()